requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
googlesearch-python==2020.0.2
idna==2.10
ordered-set==4.0.2
packaging==20.8
pyparsing==2.4.7
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
